Changes to be compilable for Win x64 ('uintptr_t' must be used instead of 'unsigned long' when casting pointer to integer)
git-svn-id: http://merlin.fit.vutbr.cz/svn/STK/trunk@290 1d61b05a-ac01-0410-9a19-bfe7c875843c

+++ src/STKLib/common.h
@@ -94,7 +94,7 @@
 #else /* We don't have any choice but to align manually */
 #  define stk_memalign(align,size,pp_orig) \
      (( *(pp_orig) = malloc( size + align - 1 )) ? \
-     (void *)( (((unsigned long)*(pp_orig)) + 15) & ~0xFUL ) : NULL )
+     (void *)( (((uintptr_t)*(pp_orig)) + 15) & ~0xFUL ) : NULL )
 #  define STK_MEMALIGN_MANUAL
 #endif
